Changes screenpeek from short to int.
git-svn-id: https://svn.eduke32.com/eduke32@593 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/duke3d.h
@@ -577,7 +577,7 @@ extern int avgfvel, avgsvel, avgavel, avghorz, avgbits, avgextbits;
 
 extern int numplayers, myconnectindex;
 extern int connecthead, connectpoint2[MAXPLAYERS];   //Player linked list variables (indeces, not connection numbers)
-extern short screenpeek;
+extern int screenpeek;
 
 extern int current_menu;
 extern int tempwallptr,animatecnt;
